reusable generate permissions
